refactor findOccurences of to be in document utilities file
