+++ src/cloudWatchLogs/registry/logStreamRegistry.ts
@@ -6,18 +6,13 @@
 import * as moment from 'moment'
 import * as vscode from 'vscode'
 import { CloudWatchLogs } from 'aws-sdk'
-import {
-    CloudWatchLogsSettings,
-    parseCloudWatchLogsUri,
-    uriToKey,
-    findOccurencesOf,
-    isLogStreamUri,
-} from '../cloudWatchLogsUtils'
+import { CloudWatchLogsSettings, parseCloudWatchLogsUri, uriToKey, isLogStreamUri } from '../cloudWatchLogsUtils'
 import { getLogger } from '../../shared/logger'
 import { INSIGHTS_TIMESTAMP_FORMAT } from '../../shared/constants'
 import { DefaultCloudWatchLogsClient } from '../../shared/clients/cloudWatchLogsClient'
 import { Timeout, waitTimeout } from '../../shared/utilities/timeoutUtils'
 import { showMessageWithCancel } from '../../shared/utilities/messages'
+import { findOccurencesOf } from '../../shared/utilities/textDocumentUtilities'
 // TODO: Add debug logging statements
 
 /**
