+++ src/shared/utilities/textDocumentUtilities.ts
@@ -8,6 +8,36 @@ import * as vscode from 'vscode'
 import { getTabSizeSetting } from './editorUtilities'
 
 /**
+ * Finds occurences of text in a document. Currently only used for highlighting cloudwatchlogs data.
+ * @param document Document to search.
+ * @param keyword Keyword to search for.
+ * @returns the ranges of each and every occurence of the keyword.
+ */
+export function findOccurencesOf(document: vscode.TextDocument, keyword: string): vscode.Range[] {
+    const ranges: vscode.Range[] = []
+    let lineNum = 0
+
+    keyword = keyword.toLowerCase()
+
+    while (lineNum < document.lineCount) {
+        const currentLine = document.lineAt(lineNum)
+        const currentLineText = currentLine.text.toLowerCase()
+        let indexOccurrence = currentLineText.indexOf(keyword, 0)
+
+        while (indexOccurrence >= 0) {
+            ranges.push(
+                new vscode.Range(
+                    new vscode.Position(lineNum, indexOccurrence),
+                    new vscode.Position(lineNum, indexOccurrence + keyword.length)
+                )
+            )
+            indexOccurrence = currentLineText.indexOf(keyword, indexOccurrence + 1)
+        }
+        lineNum += 1
+    }
+    return ranges
+}
+/**
  * If the specified document is currently open, and marked as dirty, it is saved.
  */
 export async function saveDocumentIfDirty(documentPath: string): Promise<void> {
